assets: add a simple model of sdssde30-1t00 pcb
very simple, contains only the essential bits dimensions of which play
important role in the case design due to size restrictions.

there are two std parts: regular with sketches with sketches to
construct the body, and a simple refined version for the provcase
assembly.

the regular variant also has a master sketch to aid in defining
constraints in the provcase design.