fff: improve and split the assembly into a separate file
this makes it a lot easier to work on a part and look at it in the
assembled view by simply switching tabs. the separate file links to
the original provcase.FCStd. so, need both files to work on the assembly.

it is now possibly to simply select "exploded" or "assembled" option
from a "view" property of the top level assembly std part: no need to
manually modify the asmsheet. the property config is in the same
spreadsheet.

display standoff length now correctly offsets the LCD, the board and top
heatsink while leaving the rest in place. so, it is very easy to change
the overall height.

this change also applies 4.3inch LCD updates from 0c42295 with correct
rpi board mount screw holes, since the LCD master sketch in the
assembly is a copy.