fff: replace fused parts with a single body
the fuse function in freecad didn't seem to work well when trying to go
back and update individual body parts. this commit is essentially a
remake of the plate + wall parts fuse from 6309e76 as a single body.

the reason to go back and modify individual parts was to move SSD for
more estate, to fit a 90deg connector because a straight cable doesn't
fit.

while there, added missing screws and spacers to the assembly.
this version should be good to start 3d printing.