delete fff version and rename diyplus into "nerd"
this is to align with the naming scheme we came up on the website at
https://nakamochi.io/#editions

essentially, we merged fff, aka DIY, and DIY+ which lived in diyplus
directory, into one called NERD. this is because the simplest DIY didn't
provide acceptable performance characteristics and they are both
suitable for a home-made DIY project anyway. hence, the "nerd" edition
aka "have fun playing around with it on your own".

it also makes things simpler. the other version we'll be making is KAFI.
there are no source files for it yet.