nerd/box: add a box mfr tolerance parameter
earlier, some dimensions were sized as dy + a hardcoded 0.5mm as a way
to adjust for FDM inaccuracy.

with this commit, the adjustment is now moved to a new parameter
box_mfr_tolerance so that it's easier to prepare for different tech
process: FDM, SLA, jet, etc.

the parameter is now set to 0mm because we're moving towards more
accurate processes, SLA and jet. but it is always possible to set this
back to 0.5mm or another value and generate an STL mesh file.
just watch out for freecad's topo naming problem.